Revert "Persist and restore query text, tab position and splitter direction in QueryTabComponent (#1993)"
This reverts commit d562fc0f40e0d7de3429fa3b8c70cb6242b43142.
