+++ src/Explorer/Tabs/QueryTab/QueryTabComponent.test.tsx
@@ -34,7 +34,6 @@ jest.mock("Shared/AppStatePersistenceUtility", () => ({
   AppStateComponentNames: {
     QueryCopilot: "QueryCopilot",
   },
-  readSubComponentState: jest.fn(),
 }));
 
 describe("QueryTabComponent", () => {
